chore(delete-question-use-case): add author validation

+++ src/domain/forum/application/use-cases/delete-question/delete-question.ts
@@ -4,11 +4,18 @@ import { DeleteQuestionUseCaseProps } from "./delete-question.types";
 export class DeleteQuestionUseCase {
 	constructor(private questionsRepository: QuestionsRepository) {}
 
-	async execute({ questionId }: DeleteQuestionUseCaseProps): Promise<void> {
+	async execute({
+		authorId,
+		questionId,
+	}: DeleteQuestionUseCaseProps): Promise<void> {
 		const question = await this.questionsRepository.findById(questionId);
 
 		if (!question) {
 			throw new Error("Question nof found.");
+		}
+
+		if (authorId !== question.authorId.toString()) {
+			throw new Error("Not allowed.");
 		}
 
 		await this.questionsRepository.delete(question);
